+++ .eslintrc.ts.js
@@ -8,10 +8,15 @@ module.exports = {
         ecmaVersion: 2018,
         sourceType: 'module',
     },
-    plugins: ['prettier', '@typescript-eslint'],
-    extends: ['plugin:@typescript-eslint/recommended'],
+    plugins: ['prettier', '@typescript-eslint', 'eslint-plugin-react', 'eslint-plugin-react-hooks'],
+    extends: ['plugin:@typescript-eslint/recommended', 'plugin:react/recommended', 'plugin:react-hooks/recommended'],
     rules: {
         'prettier/prettier': 'error',
         'no-unused-vars': ['error', { ignoreRestSiblings: true }],
+    },
+    settings: {
+        react: {
+            version: '17.0',
+        },
     },
 }